+++ app/src/main/java/org/linphone/ui/main/settings/viewmodel/SettingsViewModel.kt
@@ -112,6 +112,8 @@ class SettingsViewModel
         AppUtils.getString(R.string.contact_editor_last_name),
     )
     val sortContactsByValues = arrayListOf(0, 1)
+
+    val editNativeContactsInLinphone = MutableLiveData<Boolean>()
     val hideEmptyContacts = MutableLiveData<Boolean>()
 
     val ldapAvailable = MutableLiveData<Boolean>()
@@ -343,6 +345,7 @@ class SettingsViewModel
             )
 
             sortContactsBy.postValue(if (corePreferences.sortContactsByFirstName) 0 else 1)
+            editNativeContactsInLinphone.postValue(corePreferences.editNativeContactsInLinphone)
             hideEmptyContacts.postValue(corePreferences.hideContactsWithoutPhoneNumberOrSipAddress)
             presenceSubscribe.postValue(core.isFriendListSubscriptionEnabled)
 
@@ -582,6 +585,15 @@ class SettingsViewModel
     fun setContactSorting(sortingValue: Int) {
         coreContext.postOnCoreThread { core ->
             corePreferences.sortContactsByFirstName = sortingValue == 0
+        }
+    }
+
+    @UiThread
+    fun toggleEditNativeContactsInLinphone() {
+        val newValue = editNativeContactsInLinphone.value == false
+        coreContext.postOnCoreThread {
+            corePreferences.editNativeContactsInLinphone = newValue
+            editNativeContactsInLinphone.postValue(newValue)
         }
     }